backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
PyMuPDF==1.24.1
faiss-cpu==1.8.0
openai==1.23.6
sentence-transformers==2.7.0
gtts==2.5.1
deep-translator==1.11.4
pydantic==2.7.0
numpy==1.26.4
python-dotenv==1.0.1
httpx==0.27.0
scikit-learn==1.4.2
aiofiles==23.2.1
